requirements.txt
fastapi
httpx
bs4
pandas
slowapi
redis
python-dotenv
uvicorn
lxml